Modify data generator and network training.
When training the three-armed bandit task, use the reward of next trial to train the current trial. If the next trial is rewarded, the loss will be larger than 0, hence, hidden units value will be updted in backward propogation. Otherwise, the no-reward next trial won't make contribution to the training.